remove intermediate MonitoringActivity
