+++ app/src/main/java/com/nederlonder/wqms/MainActivity.java
@@ -2,12 +2,9 @@ package com.nederlonder.wqms;
 
 import android.content.ActivityNotFoundException;
 import android.content.Intent;
-import android.graphics.Rect;
 import android.net.Uri;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.util.Log;
-import android.view.MotionEvent;
 import android.view.View;
 import android.widget.Button;
 import android.widget.ImageButton;
@@ -17,16 +14,27 @@ public class MainActivity extends AppCompatActivity {
     private String TAG = "[MainActivity]";
 
     private ImageButton logoButton;
-    private Button monitoringButton;
+    private Button monitorLiveGraph;
+    private Button monitorHistoricalGraph;
     private Button aboutButton;
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        // todo set up logo redirections
         logoButton = findViewById(R.id.logo_button);
+        aboutButton = findViewById(R.id.about_button);
+        monitorLiveGraph = findViewById(R.id.button_monitor_live_graph);
+        monitorHistoricalGraph = findViewById(R.id.button_monitor_historical_graph);
+
+        setupButtons();
+    }
+
+    private void setupButtons() {
+
+        // todo set up logo redirections
 //        logoButton.setOnTouchListener(new View.OnTouchListener() {
 //            @Override
 //            public boolean onTouch(View view, MotionEvent motionEvent) {
@@ -42,14 +50,12 @@ public class MainActivity extends AppCompatActivity {
             public void onClick(View view) {
                 boolean clickWasOnRight = true;
                 if (clickWasOnRight)
-                    redirect("http://www.sonoma.edu");
+                    openWithChrome("http://www.sonoma.edu");
                 else
-                    redirect("http://www.atu.edu");
+                    openWithChrome("http://www.atu.edu");
             }
         });
 
-
-        aboutButton = findViewById(R.id.about_button);
         aboutButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
@@ -59,18 +65,22 @@ public class MainActivity extends AppCompatActivity {
             }
         });
 
-
-        monitoringButton = findViewById(R.id.monitoring_button);
-        monitoringButton.setOnClickListener(new View.OnClickListener() {
+        monitorLiveGraph.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Intent intent = new Intent( MainActivity.this, MonitoringActivity.class);
-                startActivity(intent);
+                LiveGraphActivity.start(MainActivity.this);
+            }
+        });
+
+        monitorHistoricalGraph.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                HistoricalActivity.start(MainActivity.this);
             }
         });
     }
 
-    private void redirect(String urlString) {
+    private void openWithChrome(String urlString) {
         Intent intent = new Intent(Intent.ACTION_VIEW, Uri.parse(urlString));
         intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
         intent.setPackage("com.android.chrome");
